♻️ Refactor: 코드 리팩토링 advisory lock으로 통합
