+++ src/main/java/bumblebee/xchangepass/domain/user/entity/User.java
@@ -1,9 +1,6 @@
 package bumblebee.xchangepass.domain.user.entity;
 
-import bumblebee.xchangepass.domain.cardTransaction.entity.CardTransaction;
-import bumblebee.xchangepass.domain.exchangeTransaction.entitiy.ExchangeTransaction;
 import bumblebee.xchangepass.domain.user.entity.value.*;
-import bumblebee.xchangepass.domain.wallet.transaction.entity.WalletTransaction;
 import bumblebee.xchangepass.domain.wallet.wallet.entity.Wallet;
 import jakarta.persistence.*;
 import lombok.Builder;
@@ -15,7 +12,6 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
 import static lombok.AccessLevel.PROTECTED;
 
@@ -68,18 +64,6 @@ public class User {
 
     @OneToOne(mappedBy ="user", orphanRemoval = true, cascade = CascadeType.ALL, fetch = FetchType.EAGER)
     private Wallet wallet;
-
-    @OneToMany(mappedBy = "user", orphanRemoval = true, cascade = CascadeType.ALL)
-    private List<ExchangeTransaction> exchangeTransactions;
-
-    @OneToMany(mappedBy = "user", orphanRemoval = true, cascade = CascadeType.ALL)
-    private List<CardTransaction> cardTransactions;
-
-    @OneToMany(mappedBy = "sender", orphanRemoval = true, cascade = CascadeType.ALL)
-    private List<WalletTransaction> sentWalletTransactions;
-
-    @OneToMany(mappedBy = "receiver", orphanRemoval = true, cascade = CascadeType.ALL)
-    private List<WalletTransaction> receivedWalletTransactions;
 
     @CreatedDate
     @Column(name = "user_join_date")
